moved all conftests to a single file
